+++ packages/manager/src/features/linodes/LinodesLanding/DeleteLinodeDialog.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Typography from 'src/components/core/Typography';
-import Notice from 'src/components/Notice';
+import { Notice } from 'src/components/Notice/Notice';
 import TypeToConfirmDialog from 'src/components/TypeToConfirmDialog';
 import {
   useDeleteLinodeMutation,