+++ CHANGELOG.md
@@ -8,14 +8,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added:
 - Resource links to empty state Volumes landing page #9065
-- Ability to search Linodes by IPv6 #9073
+- Resource links to empty state Firewalls landing page #9078
+- Resource links to empty state Domains landing page #9092
 - Ability download DNS zone file #9075
 
 ### Changed:
 - Removed MongoDB reference from ClusterControl description #9081
+- Highlighted Marketplace apps and button card height on empty state Linodes landing page #9083
 
 ### Fixed:
-- Allow simple search for IPv6 addresses #5088
+- Ability to search Linodes by IPv6 #9073
 
 ### Tech Stories:
 - Enable Cypress `experimentalMemoryManagement` #9076
@@ -25,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - MUI v5 Migration - `Components > TableBody` #9082
 - MUI v5 Migration - `Components > TableSortCell` #9082
 - MUI v5 Migration - `Components > TableRow` #9082
+- MUI v5 Migration - `Components > Notice` #9094
 - React Query for Linodes Landing #9062
 
 ## [2023-05-01] - v1.92.0